Issue-434: Moved UIPerson to core module
