+++ app/src/main/java/com/kirchhoff/movies/mapper/person/main/IPersonsMapper.kt
@@ -1,8 +1,8 @@
 package com.kirchhoff.movies.mapper.person.main
 
+import com.kirchhoff.movies.core.data.UIPerson
 import com.kirchhoff.movies.core.repository.Result
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
-import com.kirchhoff.movies.data.ui.main.UIPerson
 import com.kirchhoff.movies.networkdata.core.NetworkPaginated
 import com.kirchhoff.movies.networkdata.main.NetworkPerson
