+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/persons/PersonsVM.kt
@@ -1,8 +1,8 @@
 package com.kirchhoff.movies.ui.screens.main.persons
 
+import com.kirchhoff.movies.core.data.UIPerson
 import com.kirchhoff.movies.core.ui.paginated.PaginatedScreenVM
 import com.kirchhoff.movies.core.ui.paginated.UIPaginated
-import com.kirchhoff.movies.data.ui.main.UIPerson
 import com.kirchhoff.movies.repository.person.IPersonsRepository
 
 class PersonsVM(private val personRepository: IPersonsRepository) : PaginatedScreenVM<UIPaginated<UIPerson>>() {
