+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/persons/adapter/PersonVH.kt
@@ -4,10 +4,10 @@ import android.view.ViewGroup
 import android.widget.ImageView
 import android.widget.TextView
 import com.kirchhoff.movies.R
+import com.kirchhoff.movies.core.data.UIPerson
 import com.kirchhoff.movies.core.extensions.downloadAvatar
 import com.kirchhoff.movies.core.extensions.inflate
 import com.kirchhoff.movies.core.ui.recyclerview.adapter.viewholder.BaseVH
-import com.kirchhoff.movies.data.ui.main.UIPerson
 
 class PersonVH(parent: ViewGroup) : BaseVH<UIPerson>(parent.inflate(R.layout.item_person)) {
