+++ app/src/test/java/com/kirchhoff/movies/mapper/person/main/PersonsMapperTest.kt
@@ -1,7 +1,7 @@
 package com.kirchhoff.movies.mapper.person.main
 
+import com.kirchhoff.movies.core.data.UIPerson
 import com.kirchhoff.movies.core.repository.Result
-import com.kirchhoff.movies.data.ui.main.UIPerson
 import com.kirchhoff.movies.networkdata.core.NetworkPaginated
 import com.kirchhoff.movies.networkdata.main.NetworkPerson
 import com.kirchhoff.movies.utils.nextString
